# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CU Faculty Backend API
  version: "1.0.0"
  description: >
    REST API for the CU Faculty student hub. All endpoints under `/api` require a
    valid `bff.cookie` session cookie, which is validated against the Central
    University (CU) upstream on every request. The authenticated CU student id is
    used as the local user id. The only unauthenticated endpoint is `/health`.


    List endpoints are paginated and return a `{ data, total, limit, offset }`
    envelope. Pagination is controlled with the `limit` (default 20, max 100) and
    `offset` (default 0) query parameters.
security:
  - cookieAuth: []
...
paths:
...
  /api/me/socials:
    get:
      tags: [Me]
      summary: List my socials
      operationId: getMySocials
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged socials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SocialPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Me]
      summary: Create a social link
      operationId: createSocial
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SocialRequest" }
      responses:
        "201":
          description: Created social
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Social" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/me/socials/{socialId}:
    put:
      tags: [Me]
      summary: Update a social link
      operationId: updateSocial
      parameters:
        - $ref: "#/components/parameters/SocialId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SocialRequest" }
      responses:
        "200":
          description: Updated social
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Social" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Me]
      summary: Delete a social link
      operationId: deleteSocial
      parameters:
        - $ref: "#/components/parameters/SocialId"
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/me/news:
    get:
      tags: [Me]
      summary: List news I authored
      operationId: getMyNews
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged news
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewsPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/news:
    get:
      tags: [News]
      summary: List news
      operationId: listNews
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paged news
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewsPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [News]
      summary: Create a news item
      description: >-
        Accepts multipart/form-data: a JSON `data` part with the news fields and
        an optional `photo` image part uploaded in the same request.
      operationId: createNews
      requestBody:
        required: true
        description: >
          Actually sent as `multipart/form-data`: a required JSON `data` part
          matching the schema below, plus an optional `photo` image part
          (Content-Type must start with `image/`). Documented as `application/json`
          so the `data` schema renders; only the `data` part is shown here.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateNewsData" }
      responses:
        "201":
          description: Created news
          content:
            application/json:
              schema: { $ref: "#/components/schemas/News" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/news/{id}:
    get:
      tags: [News]
      summary: Get a news item
      operationId: getNews
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: News
          content:
            application/json:
              schema: { $ref: "#/components/schemas/News" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [News]
      summary: Update a news item (author only)
      description: >-
        Accepts multipart/form-data with an optional JSON `data` part (news
        fields, partial update) and an optional `photo` image part. Supplying
        `photo` replaces the current one. Either or both parts may be omitted.
      operationId: updateNews
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: false
        description: >
          Actually sent as `multipart/form-data`: an optional JSON `data` part
          matching the schema below (partial update), plus an optional `photo`
          image part (Content-Type must start with `image/`) that replaces the
          current one. Documented as `application/json` so the `data` schema
          renders; only the `data` part is shown here.
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateNewsData" }
      responses:
        "200":
          description: Updated news
          content:
            application/json:
              schema: { $ref: "#/components/schemas/News" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [News]
      summary: Delete a news item (author only)
      operationId: deleteNews
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204": { description: Deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bff.cookie
      description: CU session cookie, validated against the CU upstream on every request.
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Page size (default 20, clamped to a maximum of 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      required: false
      description: Number of items to skip (default 0).
      schema: { type: integer, minimum: 0, default: 0 }
...
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    SocialId:
      name: socialId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or rejected session cookie
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: >
        Normalized error envelope returned by every error response. "error" is a
        human-readable message (not for programmatic use); "code" is a stable,
        machine-readable identifier clients may branch on.
      properties:
        error: { type: string, example: "announcement not found" }
        code:
          type: string
          example: not_found
          enum:
            - bad_request
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - already_exists
            - invalid_reference
            - payload_too_large
            - upstream_error
            - unavailable
            - internal_error
      required: [error, code]
    PageMeta:
      type: object
      properties:
        total: { type: integer, description: Total number of matching items }
        limit: { type: integer }
        offset: { type: integer }
      required: [total, limit, offset]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        photoUrl:
          type: string
          nullable: true
          description: Presigned, time-limited download URL for the user's photo.
        firstName: { type: string, maxLength: 31 }
        lastName: { type: string, maxLength: 31 }
        bio: { type: string, nullable: true, maxLength: 255 }
        birthdate: { type: string, format: date, example: "2000-01-31" }
        speciality: { type: string, nullable: true, maxLength: 63 }
        status: { type: string, nullable: true, maxLength: 127, description: Resolved status content }
        role: { type: string, enum: [user, admin] }
      required: [id, firstName, lastName, birthdate, role]
...
    Social:
      type: object
      properties:
        id: { type: integer }
        social: { $ref: "#/components/schemas/SocialNetwork" }
        link: { type: string, maxLength: 127 }
        isPreferred: { type: boolean }
      required: [id, social, link, isPreferred]
    SocialRequest:
      type: object
      properties:
        social: { $ref: "#/components/schemas/SocialNetwork" }
        link: { type: string, maxLength: 127 }
        isPreferred: { type: boolean }
      required: [social, link, isPreferred]
...
    CreateNewsData:
      type: object
      required: [title, content]
      properties:
        title: { type: string, maxLength: 127 }
        content: { type: string, maxLength: 255 }
        publishDays: { type: integer, minimum: -32768, maximum: 32767, default: 7, description: "Defaults to 7 when <= 0" }
        isDraft: { type: boolean }
    UpdateNewsData:
      type: object
      description: Partial update; omitted fields keep their current value
      properties:
        title: { type: string, maxLength: 127 }
        content: { type: string, maxLength: 255 }
        publishDays: { type: integer, minimum: -32768, maximum: 32767, default: 7, description: "Defaults to 7 when <= 0" }
        isDraft: { type: boolean }
...
    News:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 127 }
        content: { type: string, maxLength: 255 }
        photoUrl: { type: string, nullable: true, description: Presigned download URL }
        publishDays: { type: integer, minimum: -32768, maximum: 32767 }
        isDraft: { type: boolean }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/User" }
      required: [id, title, content, publishDays, isDraft, createdAt, author]
...
    SocialNetwork:
      type: string
      enum: [vk, telegram, github]
...
    SocialPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/Social" } }
          required: [data]
...
    NewsPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            data: { type: array, items: { $ref: "#/components/schemas/News" } }
          required: [data]
...